rename Hyrax::PcdmFileFet to Hyrax::FileSet
