+++ app/services/hyrax/custom_queries/navigators/find_files.rb
@@ -23,7 +23,7 @@ module Hyrax
         delegate :resource_factory, to: :query_service
 
         # Find file ids of a given file set resource, and map to file metadata resources
-        # @param file_set [Hyrax::PcdmFileSet]
+        # @param file_set [Hyrax::FileSet]
         # @return [Array<Hyrax::FileMetadata>]
         def find_files(file_set:)
           if file_set.respond_to?(:file_ids)
@@ -31,12 +31,12 @@ module Hyrax
             query_service.custom_queries.find_many_file_metadata_by_ids(ids: file_set.file_ids)
           else
             raise ::Valkyrie::Persistence::ObjectNotFoundError,
-                  "#{file_set.internal_resource} is not a `Hydra::PcdmFileSet` implementer"
+                  "#{file_set.internal_resource} is not a `Hydra::FileSet` implementer"
           end
         end
 
         # Find original file id of a given file set resource, and map to file metadata resource
-        # @param file_set [Hyrax::PcdmFileSet]
+        # @param file_set [Hyrax::FileSet]
         # @return [Hyrax::FileMetadata]
         def find_original_file(file_set:)
           if file_set.respond_to?(:original_file_id)
@@ -44,12 +44,12 @@ module Hyrax
             query_service.custom_queries.find_file_metadata_by(id: file_set.original_file_id)
           else
             raise ::Valkyrie::Persistence::ObjectNotFoundError,
-                  "#{file_set.internal_resource} is not a `Hydra::PcdmFileSet` implementer"
+                  "#{file_set.internal_resource} is not a `Hydra::FileSet` implementer"
           end
         end
 
         # Find extracted text id of a given file set resource, and map to file metadata resource
-        # @param file_set [Hyrax::PcdmFileSet]
+        # @param file_set [Hyrax::FileSet]
         # @return [Hyrax::FileMetadata]
         def find_extracted_text(file_set:)
           if file_set.respond_to?(:extracted_text_id)
@@ -57,12 +57,12 @@ module Hyrax
             query_service.custom_queries.find_file_metadata_by(id: file_set.extracted_text_id)
           else
             raise ::Valkyrie::Persistence::ObjectNotFoundError,
-                  "#{file_set.internal_resource} is not a `Hydra::PcdmFileSet` implementer"
+                  "#{file_set.internal_resource} is not a `Hydra::FileSet` implementer"
           end
         end
 
         # Find thumbnail id of a given file set resource, and map to file metadata resource
-        # @param file_set [Hyrax::PcdmFileSet]
+        # @param file_set [Hyrax::FileSet]
         # @return [Hyrax::FileMetadata]
         def find_thumbnail(file_set:)
           if file_set.respond_to?(:thumbnail_id)
@@ -70,7 +70,7 @@ module Hyrax
             query_service.custom_queries.find_file_metadata_by(id: file_set.thumbnail_id)
           else
             raise ::Valkyrie::Persistence::ObjectNotFoundError,
-                  "#{file_set.internal_resource} is not a `Hydra::PcdmFileSet` implementer"
+                  "#{file_set.internal_resource} is not a `Hydra::FileSet` implementer"
           end
         end
       end
